broken-hospital/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
PyYAML==6.0.1
flask-swagger-ui==4.11.1
python-dotenv==1.0.0
Flask-Mail==0.9.1
SQLAlchemy==2.0.23
pytest==7.4.3
pytest-cov==4.1.0
aiosmtpd==1.4.4
aiohttp==3.8.1 